usb: dwc3-msm: Add interconnect bus voting through dt-node
Currently, the bus votes are hardcoded in the driver,
therefore having different votes for different SOCs is
not possible as the vote values will remain the same for
all. However, it is possible to do it through dt files as
it would be different for every SOC.
Implemented this by introducing a function
dwc3_msm_interconnect_vote_populate which reads the dt
file values passed and update the existing array
bus_vote_values of NOM & SVS used for bus voting.

Change-Id: If22d4e98f76e2a2e7aaebfe632e20a2225e30b3c
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>
Signed-off-by: Akash Kumar <akakum@qti.qualcomm.com>