bus: mhi: devices: Reinit completions before sending SYS_ERR command
Fix a race condition in the MHI satellite driver where the host
sends a SYS_ERR command to the ADSP and then reinitializes the
completion object before waiting for the response. If the ADSP
processes the SYS_ERR and sends the completion before the host
reinitializes it, the completion is missed, causing
wait_for_completion() to hang.

To avoid this, reinitialize the completion just after
wait_for_completion_timeout(), this ensures that any incoming
completion is properly captured.

Sequence before fix:
	+----+					+----+
	|APPS|					|ADSP|
	+----+					+----+
	  |                                       |
	SYS_ERR cmd sent------------------------->|
	  |<--------------------------------------|
	  | 					  |
	  |<------------------------ADSP process SYS_ERR and sent compeltion
	reinit_completion()                       |
	  |                                       |
	wait_for_completion()                     |
	  |                                       |

Fixes the issue by reordering the operations to avoid missing
the completion signal.

Change-Id: Ic17d89373a49907c0e5a940cca1c05737b3dba69
Signed-off-by: Vivek Pernamitta <quic_vpernami@quicinc.com>