modules.list.msm: Move qpnp-power-on.ko to first stage
In change 5954080 ("soc: qcom: Add support of reboot reason driver"),
the qcom-reboot-reason driver was updated to call
qpnp_pon_set_restart_reason() as a fallback when NVMEM cell support
is not present.

Since this driver is part of the first stage, but qpnp-power-on was
set to load in the second stage on some targets, the build failed
with unknown symbol errors during linking.

To resolve this, qpnp-power-on is now enabled in the first stage on
applicable targets, ensuring the symbol is available at the time of
use.

Change-Id: I0857d90ba939da7a684d4b1690cf72e018f427e3
Signed-off-by: Mohammed Aakib Patel <quic_maiyubpa@quicinc.com>